Revert "More unsigned ints because of undefined integer overflow behaviour"
This reverts commit 623223354a2b103443e4554cddc218d8dce2c3d2.

+++ src/engine/client/sound.cpp
@@ -260,8 +260,8 @@ static void Mix(short *pFinalOut, unsigned Frames)
 		for(unsigned i = 0; i < Frames; i++)
 		{
 			int j = i<<1;
-			unsigned int vl = (((unsigned int) m_pMixBuffer[j]*MasterVol)/101)>>8;
-			unsigned int vr = (((unsigned int) m_pMixBuffer[j+1]*MasterVol)/101)>>8;
+			int vl = ((m_pMixBuffer[j]*MasterVol)/101)>>8;
+			int vr = ((m_pMixBuffer[j+1]*MasterVol)/101)>>8;
 
 			pFinalOut[j] = Int2Short(vl);
 			pFinalOut[j+1] = Int2Short(vr);
